Update default container path to match the jfrog one.
For customers who have there own repo it can be overriden however the way that the path is currently works is not compatible with the if conditions.  By changing the JFROG to match defalt path the base JFROG can be overidden as it matches the public view of the world.